Optimize the core
The speed hack options added some overhead, let's remove that overhead.
:P

The IO-Read hack was too invasive, remove it as it slowed down the core
with it disabled. Optimized the bus slowdown hack so it doesn't insert
any actual runtime overhead for existing anymore. The hack no longer is
exactly half the bus speed, as fast ops aren't slowed down anymore,
only stuff that are configurable in wait state i/o are.